# src/api/openapi.yml
openapi: 3.0.0
info:
  version: 0.1.0
  title: AgentENV E2B-compatible API
servers:
  - url: /
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
    # AdminApiKeyAuth / AdminTeamAuth: alphabetical names ensure token
    # validation runs before team context population.
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    paginationNextToken:
      name: nextToken
      in: query
      description: Cursor to start the list from
      required: false
      schema:
        type: string
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    SandboxState:
      type: string
      description: State of the sandbox
      enum:
        - running
        - paused
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error
...
paths:
...
  /sandboxes:
    get:
      summary: List running sandboxes
      description: List all running sandboxes. Use GET /v2/sandboxes instead.
      deprecated: true
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - name: metadata
          in: query
          description: Metadata query used to filter the sandboxes (e.g. "user=abc&app=prod"). Each key and values must be URL encoded.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned all running sandboxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListedSandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
    post:
      summary: Create sandbox
      description: >
        Create a sandbox from an existing template/snapshot. Use
        POST /sandboxes-cold to create a sandbox directly from an external OCI
        image.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSandbox"
      responses:
        "201":
          description: The sandbox was created successfully
          headers:
            x-agentenv-sandbox-id:
              description: Identifier of the created sandbox for proxy routing.
              schema:
                type: string
            x-agentenv-execution-id:
              description: >-
                The incarnation this sandbox is running under. The routing
                projection is written under it, and a write that cannot name
                one is refused by incarnation arbitration rather than
                overwriting a record that names a different run.
              schema:
                type: string
            x-agentenv-projection-ttl-secs:
              description: >-
                How long this sandbox's routing projection should live, in
                whole seconds, as derived from the node's sandbox lifetime
                ceiling. Absent or 0 means the receiver should use its own
                default TTL; it never means the record should not expire.
              schema:
                type: integer
                format: int64
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
        "503":
          description: No node can take a sandbox with these network rules right now
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /sandboxes-cold:
    post:
      description: >
        Create a sandbox directly from an explicit external OCI image
        reference. Cold starts may pull and convert OCI layers on a cache miss
        and can take tens of seconds; clients should use a long HTTP timeout
        for this path.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewColdSandbox"
      responses:
        "201":
          description: The sandbox was created successfully
          headers:
            x-agentenv-sandbox-id:
              description: Identifier of the created sandbox for proxy routing.
              schema:
                type: string
            x-agentenv-execution-id:
              description: >-
                The incarnation this sandbox is running under. The routing
                projection is written under it, and a write that cannot name
                one is refused by incarnation arbitration rather than
                overwriting a record that names a different run.
              schema:
                type: string
            x-agentenv-projection-ttl-secs:
              description: >-
                How long this sandbox's routing projection should live, in
                whole seconds, as derived from the node's sandbox lifetime
                ceiling. Absent or 0 means the receiver should use its own
                default TTL; it never means the record should not expire.
              schema:
                type: integer
                format: int64
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
        "503":
          description: No node can take a sandbox with these network rules right now
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v2/sandboxes:
    get:
      summary: List sandboxes (v2)
      description: List all sandboxes
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - name: metadata
          in: query
          description: Metadata query used to filter the sandboxes (e.g. "user=abc&app=prod"). Each key and values must be URL encoded.
          required: false
          schema:
            type: string
        - name: state
          in: query
          description: Filter sandboxes by one or more states
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/SandboxState"
          style: form
          explode: false
        - $ref: "#/components/parameters/paginationNextToken"
        - $ref: "#/components/parameters/paginationLimit"
      responses:
        "200":
          description: Successfully returned all running sandboxes
          headers:
            x-next-token:
              description: Pagination token for the next page. Returned only when more results are available.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListedSandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
  /sandboxes/{sandboxID}:
    get:
      summary: Sandbox
      description: Get a sandbox by id
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      responses:
        "200":
          description: Successfully returned the sandbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxDetail"
        "404":
          $ref: "#/components/responses/404"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"

    delete:
      summary: Kill sandbox
      description: Kill a sandbox
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      responses:
        "204":
          description: The sandbox was killed successfully
        "404":
          $ref: "#/components/responses/404"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"

  # TODO: Pause and resume might be exposed as POST /sandboxes/{sandboxID}/snapshot and then POST /sandboxes with specified snapshotting setup
...
  /secrets:
    get:
      summary: List secrets
      description: List the deployment's secrets. No response carries a secret value.
      tags: [secrets]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/paginationNextToken"
        - $ref: "#/components/parameters/paginationLimit"
      responses:
        "200":
          description: Successfully listed the secrets
          headers:
            X-Next-Token:
              description: Cursor to fetch the next page of results, if more exist
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Secret"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
        "503":
          description: No secrets store is configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a secret
      description: Create a secret by storing its first version. The response carries metadata only.
      tags: [secrets]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSecret"
      responses:
        "201":
          description: Successfully created the secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
        "503":
          description: No secrets store is configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /secrets/{secretID}:
    get:
      summary: Get a secret
      description: Get one secret's metadata, selected by identifier or name.
      tags: [secrets]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/secretID"
      responses:
        "200":
          description: Successfully retrieved the secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          description: No secrets store is configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Update a secret
      description: Replace the secret's value by appending a new version. The response carries metadata only.
      tags: [secrets]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/secretID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretUpdate"
      responses:
        "200":
          description: Successfully updated the secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          description: No secrets store is configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a secret
      description: Delete the secret and every version of its value. Sandboxes whose rules reference it fail at egress from then on.
      tags: [secrets]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/secretID"
      responses:
        "204":
          description: Successfully deleted the secret
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          description: No secrets store is configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /snapshots:
    get:
      summary: List snapshots
      description: List all snapshots for the team
      tags: [snapshots]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - name: sandboxID
          in: query
          required: false
          schema:
            type: string
            description: Filter snapshots by source sandbox ID
        - name: name
          in: query
          description: Filter snapshots by name or ID, optionally tag-qualified (e.g. "my-snapshot", "my-team/my-snapshot" or "my-snapshot:v1").
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/paginationLimit"
        - $ref: "#/components/parameters/paginationNextToken"
      responses:
        "200":
          description: Successfully returned snapshots
          headers:
            x-next-token:
              description: Pagination token for the next page. Returned only when more results are available.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotInfo"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
...
  /v3/templates:
    post:
      summary: Create template (v3)
      description: Create a new template
      tags: [templates]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateBuildRequestV3"

      responses:
        "202":
          description: The build was requested successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateRequestResponseV3"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
...
  /nodes:
    get:
      summary: List nodes
      description: List all nodes
      tags: [admin]
      security:
        - AdminApiKeyAuth: []
      parameters:
        - in: query
          name: clusterID
          description: Identifier of the cluster
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully returned all nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
...